#64: Move RolePrivilege to privilege package
