+++ src/main/java/org/codenergic/theskeleton/privilege/impl/PrivilegeServiceImpl.java
@@ -15,9 +15,16 @@
  */
 package org.codenergic.theskeleton.privilege.impl;
 
+import java.util.Set;
+import java.util.stream.Collectors;
+
 import org.codenergic.theskeleton.privilege.PrivilegeEntity;
 import org.codenergic.theskeleton.privilege.PrivilegeRepository;
 import org.codenergic.theskeleton.privilege.PrivilegeService;
+import org.codenergic.theskeleton.privilege.RolePrivilegeEntity;
+import org.codenergic.theskeleton.privilege.RolePrivilegeRepository;
+import org.codenergic.theskeleton.role.RoleEntity;
+import org.codenergic.theskeleton.role.RoleRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -26,15 +33,22 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 public class PrivilegeServiceImpl implements PrivilegeService {
-	private PrivilegeRepository privilegeRepository;
+	private final PrivilegeRepository privilegeRepository;
+	private final RoleRepository roleRepository;
+	private final RolePrivilegeRepository rolePrivilegeRepository;
 
-	public PrivilegeServiceImpl(PrivilegeRepository privilegeRepository) {
+	public PrivilegeServiceImpl(PrivilegeRepository privilegeRepository, RoleRepository roleRepository, RolePrivilegeRepository rolePrivilegeRepository) {
 		this.privilegeRepository = privilegeRepository;
+		this.roleRepository = roleRepository;
+		this.rolePrivilegeRepository = rolePrivilegeRepository;
 	}
 
 	@Override
-	public PrivilegeEntity findPrivilegeByName(String name) {
-		return privilegeRepository.findByName(name);
+	@Transactional
+	public RoleEntity addPrivilegeToRole(String code, String privilegeName) {
+		RoleEntity role = roleRepository.findByCode(code);
+		PrivilegeEntity privilege = privilegeRepository.findByName(privilegeName);
+		return rolePrivilegeRepository.save(new RolePrivilegeEntity(role, privilege)).getRole();
 	}
 
 	@Override
@@ -49,12 +63,32 @@ public class PrivilegeServiceImpl implements PrivilegeService {
 	}
 
 	@Override
-	public Page<PrivilegeEntity> findPrivileges(Pageable pageable) {
-		return privilegeRepository.findAll(pageable);
+	public PrivilegeEntity findPrivilegeByName(String name) {
+		return privilegeRepository.findByName(name);
 	}
 
 	@Override
 	public Page<PrivilegeEntity> findPrivileges(String keyword, Pageable pageable) {
 		return privilegeRepository.findByNameOrDescriptionStartsWith(keyword, pageable);
+	}
+
+	@Override
+	public Page<PrivilegeEntity> findPrivileges(Pageable pageable) {
+		return privilegeRepository.findAll(pageable);
+	}
+
+	@Override
+	public Set<PrivilegeEntity> findPrivilegesByRoleCode(String code) {
+		return rolePrivilegeRepository.findByRoleCode(code).stream()
+			.map(RolePrivilegeEntity::getPrivilege)
+			.collect(Collectors.toSet());
+	}
+
+	@Override
+	@Transactional
+	public RoleEntity removePrivilegeFromRole(String code, String privilegeName) {
+		RolePrivilegeEntity userRole = rolePrivilegeRepository.findByRoleCodeAndPrivilegeName(code, privilegeName);
+		rolePrivilegeRepository.delete(userRole);
+		return roleRepository.findByCode(code);
 	}
 }
