+++ src/main/java/org/codenergic/theskeleton/user/UserAdminService.java
@@ -22,7 +22,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
 import org.codenergic.theskeleton.role.RoleEntity;
-import org.codenergic.theskeleton.role.RolePrivilegeRepository;
+import org.codenergic.theskeleton.privilege.RolePrivilegeRepository;
 import org.codenergic.theskeleton.role.RoleRepository;
 import org.codenergic.theskeleton.user.impl.UserServiceImpl;
 import org.springframework.data.domain.Page;
