+++ src/test/java/org/codenergic/theskeleton/privilege/PrivilegeServiceTest.java
@@ -15,18 +15,14 @@
  */
 package org.codenergic.theskeleton.privilege;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
 import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.UUID;
 
-import org.codenergic.theskeleton.privilege.PrivilegeEntity;
-import org.codenergic.theskeleton.privilege.PrivilegeRepository;
-import org.codenergic.theskeleton.privilege.PrivilegeService;
 import org.codenergic.theskeleton.privilege.impl.PrivilegeServiceImpl;
+import org.codenergic.theskeleton.role.RoleEntity;
+import org.codenergic.theskeleton.role.RoleRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -36,15 +32,25 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
 public class PrivilegeServiceTest {
 	private PrivilegeService privilegeService;
 	@Mock
 	private PrivilegeRepository privilegeRepository;
+	@Mock
+	private RoleRepository roleRepository;
+	@Mock
+	private RolePrivilegeRepository rolePrivilegeRepository;
 
 	@Before
 	public void init() {
 		MockitoAnnotations.initMocks(this);
-		this.privilegeService = new PrivilegeServiceImpl(privilegeRepository);
+		this.privilegeService = new PrivilegeServiceImpl(privilegeRepository, roleRepository, rolePrivilegeRepository);
 	}
 
 	@Test
@@ -79,5 +85,40 @@ public class PrivilegeServiceTest {
 		when(privilegeRepository.findAll(any(Pageable.class))).thenReturn(page);
 		assertThat(privilegeService.findPrivileges(new PageRequest(1, 10))).isEqualTo(page);
 		verify(privilegeRepository).findAll(any(Pageable.class));
+	}
+
+	@Test
+	public void testAddPrivilegeToRole() {
+		RoleEntity role = new RoleEntity()
+			.setId(UUID.randomUUID().toString())
+			.setCode("role");
+		PrivilegeEntity privilege = new PrivilegeEntity()
+			.setId(UUID.randomUUID().toString())
+			.setName("privilege");
+		RolePrivilegeEntity result = new RolePrivilegeEntity(role, privilege);
+		result.setId(UUID.randomUUID().toString());
+		when(roleRepository.findByCode("role")).thenReturn(role);
+		when(privilegeRepository.findByName("privilege")).thenReturn(privilege);
+		when(rolePrivilegeRepository.save(any(RolePrivilegeEntity.class))).thenReturn(result);
+		assertThat(privilegeService.addPrivilegeToRole("role", "privilege")).isEqualTo(role);
+		verify(roleRepository).findByCode("role");
+		verify(privilegeRepository).findByName("privilege");
+		verify(rolePrivilegeRepository).save(any(RolePrivilegeEntity.class));
+	}
+
+	@Test
+	public void testRemovePrivilegeFromRole() {
+		privilegeService.removePrivilegeFromRole("", "");
+	}
+
+	@Test
+	public void testFindPrivilegesByRoleCode() {
+		Set<RolePrivilegeEntity> dbResult =
+			new HashSet<>(Arrays.asList(new RolePrivilegeEntity().setPrivilege(new PrivilegeEntity().setName("privilege"))));
+		when(rolePrivilegeRepository.findByRoleCode("role")).thenReturn(dbResult);
+		Set<PrivilegeEntity> result = privilegeService.findPrivilegesByRoleCode("role");
+		assertThat(result.size()).isEqualTo(1);
+		assertThat(result.iterator().next()).isEqualTo(dbResult.iterator().next().getPrivilege());
+		verify(rolePrivilegeRepository).findByRoleCode("role");
 	}
 }
