+++ src/test/java/org/codenergic/theskeleton/role/RoleRestControllerTest.java
@@ -15,31 +15,12 @@
  */
 package org.codenergic.theskeleton.role;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import org.codenergic.theskeleton.core.test.EnableRestDocs;
 import org.codenergic.theskeleton.core.test.InjectUserDetailsService;
-import org.codenergic.theskeleton.privilege.PrivilegeEntity;
-import org.codenergic.theskeleton.privilege.PrivilegeRestData;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -55,6 +36,20 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @RunWith(SpringRunner.class)
 @EnableSpringDataWebSupport
@@ -193,64 +188,5 @@ public class RoleRestControllerTest {
 					.contentType(MediaType.APPLICATION_JSON))
 				.andExpect(status().isOk())
 				.andDo(document("role-delete"));
-	}
-
-	@Test
-	public void testAddPrivilegeToRole() throws Exception {
-		final RoleEntity role = new RoleEntity()
-				.setId("role123")
-				.setCode("role123");
-		when(roleService.addPrivilegeToRole("role123", "privilege123")).thenReturn(role);
-		ResultActions resultActions = mockMvc.perform(put("/api/roles/role123/privileges")
-					.content("{\"privilege\": \"privilege123\"}")
-					.contentType(MediaType.APPLICATION_JSON))
-				.andExpect(status().isOk())
-				.andDo(document("role-privilege-create"));
-		MockHttpServletResponse response = resultActions
-				.andReturn()
-				.getResponse();
-		verify(roleService).addPrivilegeToRole("role123", "privilege123");
-		assertThat(response.getContentAsByteArray())
-				.isEqualTo(objectMapper.writeValueAsBytes(RoleRestData.builder(role).build()));
-	}
-
-	@Test
-	public void testFindPrivilegesByRoleCode() throws Exception {
-		final PrivilegeEntity privilege = new PrivilegeEntity()
-				.setId("privilege123")
-				.setName("user_list_read");
-		final Set<PrivilegeEntity> privileges = new HashSet<>(Arrays.asList(privilege));
-		final Set<PrivilegeRestData> expected = privileges.stream()
-				.map(p -> PrivilegeRestData.builder(p).build())
-				.collect(Collectors.toSet());
-		when(roleService.findPrivilegesByRoleCode("role123")).thenReturn(privileges);
-		ResultActions resultActions = mockMvc.perform(get("/api/roles/role123/privileges")
-					.contentType(MediaType.APPLICATION_JSON))
-				.andExpect(status().isOk())
-				.andDo(document("role-privilege-read"));
-		MockHttpServletResponse response = resultActions
-				.andReturn()
-				.getResponse();
-		verify(roleService).findPrivilegesByRoleCode("role123");
-		assertThat(response.getContentAsByteArray()).isEqualTo(objectMapper.writeValueAsBytes(expected));
-	}
-
-	@Test
-	public void testRemovePrivilegeFromRole() throws Exception {
-		final RoleEntity role = new RoleEntity()
-				.setId("role123")
-				.setCode("role123");
-		when(roleService.removePrivilegeFromRole("role123", "privilege123")).thenReturn(role);
-		ResultActions resultActions = mockMvc.perform(delete("/api/roles/role123/privileges")
-					.content("{\"privilege\": \"privilege123\"}")
-					.contentType(MediaType.APPLICATION_JSON))
-				.andExpect(status().isOk())
-				.andDo(document("role-privilege-delete"));
-		MockHttpServletResponse response = resultActions
-				.andReturn()
-				.getResponse();
-		verify(roleService).removePrivilegeFromRole("role123", "privilege123");
-		assertThat(response.getContentAsByteArray())
-				.isEqualTo(objectMapper.writeValueAsBytes(RoleRestData.builder(role).build()));
 	}
 }
