+++ src/test/java/org/codenergic/theskeleton/role/RoleServiceTest.java
@@ -15,8 +15,10 @@
  */
 package org.codenergic.theskeleton.role;
 
-import org.codenergic.theskeleton.privilege.PrivilegeEntity;
-import org.codenergic.theskeleton.privilege.PrivilegeRepository;
+import java.util.Arrays;
+import java.util.UUID;
+
+import org.codenergic.theskeleton.role.impl.RoleServiceImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -26,13 +28,10 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.UUID;
-
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -40,15 +39,11 @@ public class RoleServiceTest {
 	private RoleService roleService;
 	@Mock
 	private RoleRepository roleRepository;
-	@Mock
-	private PrivilegeRepository privilegeRepository;
-	@Mock
-	private RolePrivilegeRepository rolePrivilegeRepository;
 
 	@Before
 	public void init() {
 		MockitoAnnotations.initMocks(this);
-		this.roleService = RoleService.newInstance(roleRepository, privilegeRepository, rolePrivilegeRepository);
+		this.roleService = new RoleServiceImpl(roleRepository);
 	}
 
 	@Test
@@ -118,51 +113,5 @@ public class RoleServiceTest {
 		roleService.deleteRole("123");
 		verify(roleRepository).findByCode("123");
 		verify(roleRepository).delete(input);
-	}
-
-	@Test
-	public void testAddPrivilegeToRole() {
-		RoleEntity role = new RoleEntity()
-				.setId(UUID.randomUUID().toString())
-				.setCode("role");
-		PrivilegeEntity privilege = new PrivilegeEntity()
-				.setId(UUID.randomUUID().toString())
-				.setName("privilege");
-		RolePrivilegeEntity result = new RolePrivilegeEntity(role, privilege);
-		result.setId(UUID.randomUUID().toString());
-		when(roleRepository.findByCode("role")).thenReturn(role);
-		when(privilegeRepository.findByName("privilege")).thenReturn(privilege);
-		when(rolePrivilegeRepository.save(any(RolePrivilegeEntity.class))).thenReturn(result);
-		assertThat(roleService.addPrivilegeToRole("role", "privilege")).isEqualTo(role);
-		verify(roleRepository).findByCode("role");
-		verify(privilegeRepository).findByName("privilege");
-		verify(rolePrivilegeRepository).save(any(RolePrivilegeEntity.class));
-	}
-
-	@Test
-	public void testRemovePrivilegeFromRole() {
-		roleService.removePrivilegeFromRole("", "");
-	}
-
-	@Test
-	public void testFindPrivilegesByRoleCode() {
-		Set<RolePrivilegeEntity> dbResult =
-				new HashSet<>(Arrays.asList(new RolePrivilegeEntity().setPrivilege(new PrivilegeEntity().setName("privilege"))));
-		when(rolePrivilegeRepository.findByRoleCode("role")).thenReturn(dbResult);
-		Set<PrivilegeEntity> result = roleService.findPrivilegesByRoleCode("role");
-		assertThat(result.size()).isEqualTo(1);
-		assertThat(result.iterator().next()).isEqualTo(dbResult.iterator().next().getPrivilege());
-		verify(rolePrivilegeRepository).findByRoleCode("role");
-	}
-
-	@Test
-	public void testFindRolesByPrivilegeName() {
-		Set<RolePrivilegeEntity> dbResult =
-				new HashSet<>(Arrays.asList(new RolePrivilegeEntity().setRole(new RoleEntity().setCode("role"))));
-		when(rolePrivilegeRepository.findByPrivilegeName("privilege")).thenReturn(dbResult);
-		Set<RoleEntity> result = roleService.findRolesByPrivilegeName("privilege");
-		assertThat(result.size()).isEqualTo(1);
-		assertThat(result.iterator().next()).isEqualTo(dbResult.iterator().next().getRole());
-		verify(rolePrivilegeRepository).findByPrivilegeName("privilege");
 	}
 }
